% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_generator.R
\name{arrayGen}
\alias{arrayGen}
\title{Generate array for slurm arra job}
\usage{
arrayGen(group_mode, n_inst, d_inst, member_param, samps, group, step, skip)
}
\arguments{
\item{group_mode}{standard or overlap}

\item{n_inst}{single n}

\item{d_inst}{single d}

\item{member_param}{single p or s}

\item{samps}{how many samples will we process}

\item{group}{how many indices to process per task ID}

\item{step}{step to get around slurm's 50k constraint}

\item{skip}{how many graphs (presumably done) to skip}
}
\value{
Set of scripts
}
\description{
Relies on triage job to count from 0 onwards. Meant for R. If python, need to deal
with fact that indices start at 0.
}
