% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{blockSizes_gen}
\alias{blockSizes_gen}
\title{Produce block sizes}
\usage{
blockSizes_gen(n, p, s)
}
\arguments{
\item{n}{num of vertices}

\item{p}{cluster sizes}

\item{s}{overlap}
}
\value{
whether the graph is viable
}
\description{
Produce block sizes
}
