% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{blockSizes_gen_norm}
\alias{blockSizes_gen_norm}
\title{Produce block sizes (normalized)}
\usage{
blockSizes_gen_norm(blockSize)
}
\arguments{
\item{blockSize}{vector with block sizes}
}
\value{
normalized block sizes
}
\description{
Produce block sizes (normalized)
}
