% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_job.R
\name{get_output_full}
\alias{get_output_full}
\title{Full output generation}
\usage{
get_output_full(param_mode, group_mode)
}
\arguments{
\item{param_mode}{"semi" or "non"}

\item{group_mode}{"std" or "overlap"}
}
\value{
Full output
}
\description{
Full output generation
}
