% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_job.R
\name{get_output_meet_full}
\alias{get_output_meet_full}
\title{All observations with meeting columns}
\usage{
get_output_meet_full(param_mode, group_mode)
}
\arguments{
\item{param_mode}{"semi" or "non"}

\item{group_mode}{"std" or "overlap"}
}
\value{
Non-summarized output (with meeting columns)
}
\description{
All observations with meeting columns
}
