% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_job.R
\name{get_output_sum}
\alias{get_output_sum}
\title{Summary of full output}
\usage{
get_output_sum(param_mode, group_mode)
}
\arguments{
\item{param_mode}{"semi" or "non"}

\item{group_mode}{"std" or "overlap"}
}
\value{
Full summarized output
}
\description{
Summary of full output
}
