% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_job.R
\name{get_output_sum_meet}
\alias{get_output_sum_meet}
\title{Summary of full output for meeting}
\usage{
get_output_sum_meet(output_full_meet, param_mode, group_mode)
}
\arguments{
\item{output_full_meet}{full results (from cluster)}

\item{param_mode}{"semi" or "non"}

\item{group_mode}{"std" or "overlap"}
}
\value{
Full summarized output
}
\description{
Summary of full output for meeting
}
