% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_job.R
\name{get_samps_processed}
\alias{get_samps_processed}
\title{Summary of indices processed}
\usage{
get_samps_processed(param_mode, group_mode)
}
\arguments{
\item{param_mode}{"semi" or "non"}

\item{group_mode}{"std" or "overlap"}
}
\value{
Tally of indices procced
}
\description{
Summary of indices processed
}
