% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{graph_generator}
\alias{graph_generator}
\title{Generate graph from SBM}
\usage{
graph_generator(graphNum)
}
\arguments{
\item{graphNum}{graph number}
}
\value{
graph edge list
}
\description{
Generate graph from SBM
}
