% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{graph_stats}
\alias{graph_stats}
\title{Produce stats for a given graphs}
\usage{
graph_stats(params)
}
\arguments{
\item{params}{row of parameter matrix}
}
\value{
graph summary df
}
\description{
Produce stats for a given graphs
}
