% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{matrixGenerator}
\alias{matrixGenerator}
\title{Generate M matrix (see Lelarge)}
\usage{
matrixGenerator(params)
}
\arguments{
\item{params}{Full set of parameters}
}
\value{
Matrix M
}
\description{
Notice that some entries may be illegal. This issue is taken care of in script below.
}
