% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{matrixGenerator_overlap}
\alias{matrixGenerator_overlap}
\title{Generate M matrix for overlap}
\usage{
matrixGenerator_overlap(params)
}
\arguments{
\item{params}{Full set of parameters}
}
\value{
Matrix M
}
\description{
Generate M matrix for overlap
}
