% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_stats.R
\name{MSE}
\alias{MSE}
\title{Mean squared error
NOTE: Modify when using the latest python notebook (i.e., when pv for s = 0 actually has only 2 cols).}
\usage{
MSE(pv, Z)
}
\arguments{
\item{pv}{sufficient stats}

\item{Z}{true labels}
}
\value{
MSE
}
\description{
Mean squared error
NOTE: Modify when using the latest python notebook (i.e., when pv for s = 0 actually has only 2 cols).
}
