% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_stats.R
\name{NMI}
\alias{NMI}
\title{Normalized mutual information}
\usage{
NMI(X, Y)
}
\arguments{
\item{X}{True labels}

\item{Y}{Predicted labels}
}
\value{
NMI score
}
\description{
Sources:
 https://www.researchgate.net/post/How_can_i_calculate_Mutual_Information_theory_from_a_simple_dataset
 https://en.wikipedia.org/wiki/Mutual_information
}
