% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{nullFinder}
\alias{nullFinder}
\title{Determine whether a prefernce matrix is viable}
\usage{
nullFinder(M)
}
\arguments{
\item{M}{preference matrix}
}
\value{
whether the graph is viable
}
\description{
Determine whether a prefernce matrix is viable
}
