% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{outputGen}
\alias{outputGen}
\title{Produce (single-row) algorithm output as txt file}
\usage{
outputGen(file_in, pyJobOut)
}
\arguments{
\item{file_in}{file name of graph}

\item{pyJobOut}{raw output from python}
}
\description{
Produce (single-row) algorithm output as txt file
}
