% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{outputParser}
\alias{outputParser}
\title{Parse single-row peixoto output}
\usage{
outputParser(filePath, group_mode)
}
\arguments{
\item{filePath}{path to output file}
}
\value{
A one-row data frame
}
\description{
Parse single-row peixoto output
}
