% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{paramGenerator}
\alias{paramGenerator}
\title{Generate entries of prefernce matrix}
\usage{
paramGenerator(n, d, p, lambda, sign)
}
\arguments{
\item{n}{Number of vertices (i.e., nodes) (integer)}

\item{d}{Average vertex degree (integer)}

\item{p}{Proportion of vertices belonging to the 2nd group (double)}

\item{lambda}{Relevant eigenvalue of matrix R (see Lelarge) (double)}

\item{sign}{Sign (i.e., +/-) associated with b value calculation (character)}
}
\value{
A one-row data.frame with the full parameter set
}
\description{
Creates the full parameter list from a given n, d, p, lambda, and sign (i.e., +/-).
Notice that some values are illegal entries (i.e., values lead to probabilities that
are negative or greater than 1). This issue is taken care of later.
}
