% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_generator.R
\name{script_generator}
\alias{script_generator}
\title{Generate slurm scripts}
\usage{
script_generator(param_mode, group_mode, HARVEST_NITER, n, d, member_param,
  memReq, arrayStr, group, step)
}
\arguments{
\item{param_mode}{semi or non}

\item{group_mode}{std or overlap}

\item{n}{single value}

\item{d}{single value}

\item{memReq}{single value}

\item{arrayStr}{string of task indices}

\item{group}{size of agglomeration}

\item{step}{size of step}

\item{p}{single value}
}
\description{
Generate slurm scripts
}
