% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_generator.R
\name{scripts_all}
\alias{scripts_all}
\title{Batch generate all scripts}
\usage{
scripts_all(param_mode, group_mode, HARVEST_NITER, samps = 100, group = 1,
  step = 0, skip = 0)
}
\arguments{
\item{param_mode}{semi or non}

\item{group_mode}{std or overlap}

\item{HARVEST_NITER}{number of harvest iterations}

\item{samps}{number of samps to process}

\item{group}{how many indices to process per task}

\item{step}{step to get around slurm's 50k constraint}

\item{skip}{number of graphs to skip}
}
\value{
Set of scripts
}
\description{
Batch generate all scripts
}
