% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{separation_params}
\alias{separation_params}
\title{Produce separation (i.e., c) params}
\usage{
separation_params(c, d, p)
}
\arguments{
\item{c}{sepration}

\item{d}{avg degree of graph}

\item{p}{cluster sizes}
}
\value{
entries of M matrix
}
\description{
Produce separation (i.e., c) params
}
