% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_stats.R
\name{Z_gen}
\alias{Z_gen}
\title{Generate true labels}
\usage{
Z_gen(n, p, s)
}
\arguments{
\item{n}{number of nodes}

\item{p}{proportion of nodes in 1st cluster}

\item{s}{degree of overlap}
}
\description{
Generate true labels
}
